requirements.txt
streamlit>=1.33.0
pandas>=2.2.0
python-dotenv>=1.0.1
python-telegram-bot==20.6
requests>=2.31.0
openpyxl>=3.1.2
watchdog>=3.0.0
mysql-connector-python==8.4.0
